// F1UI/F1-UI-Angular/src/app/api/model/point.ts
/**
 * F1Website
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Point { 
    driverId?: number;
    raceId?: number;
    points?: number | null;
    position?: number | null;
}