// F1UI/F1-UI-Angular/src/app/api/model/driverTeamDto.ts
/**
 * F1Website
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DriverTeamDto { 
    id?: number;
    teamId?: number;
    name?: string | null;
    country?: string | null;
    raceNumber?: number | null;
    team?: string | null;
}